Refactor formatters to make code cleaner
